requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.16.2
keras==3.3.3
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
groq==0.9.0
google-generativeai==0.7.2
python-dotenv==1.0.1
scikit-image==0.22.0
matplotlib==3.8.4
gdown==5.2.0